Refactor the command dir and fix bad path in Makefile
